Some refactor and fix the enum usage in tests
