+++ lib/pbench/test/unit/server/query_apis/commons.py
@@ -7,7 +7,7 @@ from http import HTTPStatus
 from typing import AnyStr, Type
 
 from pbench.server.api.resources.query_apis import ElasticBase, JSON, ParamType
-from pbench.test.unit.server.conftest import HeaderTypes
+from pbench.test.unit.server.headertypes import HeaderTypes
 
 
 class Commons:
@@ -218,7 +218,7 @@ class Commons:
         if not self.empty_es_response_payload or not self.elastic_endpoint:
             pytest.skip("skipping " + self.test_empty_query.__name__)
         expected_status = HTTPStatus.OK
-        if build_auth_header["header_param"] not in HeaderTypes.valid_headers():
+        if not HeaderTypes.is_valid(build_auth_header["header_param"]):
             expected_status = HTTPStatus.FORBIDDEN
 
         index = self.build_index(
